refactor: use SSR rendering and remove the API endpoint for initial stats
